% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp}
\alias{esp}
\title{13 PIACC variables for Spain}
\format{A data frame with 1991 rows and 13 variables:
\itemize{
  \item CNTRYID: Country ID (ISO 3166, numeric)
  \item SEQID: Sequential ID (randomly derived)
  \item GENDER_R: Person resolved gender from BQ and QC check (derived)
  \item EARNHR: Hourly earnings excluding bonuses for wage and salary earners (derived)
  \item EARNMTH: Monthly earnings excluding bonuses for wage and salary earners (derived)
  \item D_Q16b: Current work - Earnings - Gross pay
  \item EDCAT7: Highest level of formal education obtained (7 categories - derived)
  \item YRSQUAL: Highest level of education obtained imputed into years of education (derived)
  \item C_Q09_C: Current status/work history - Years of paid work during lifetime (top-coded at 47)
  \item C_D05: Current status/work history - Employment status (DERIVED BY CAPI)
  \item J_Q02a: Background - Living with spouse or partner
  \item J_Q04a: Background - Born in country
  \item PVNUM1: Numeracy scale score - Plausible value 1
}}
\usage{
esp
}
\description{
PIACC es el acr칩nimo de
Programa para la Evaluaci칩n Internacional de las
Competencias de los adultos (PIAAC) de la OCDE.
Informaci칩n detallada sobre el programa PIAAC puede encontrarse en su p치gina web
http://www.oecd.org/skills/piaac/
}
\keyword{datasets}

