% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fun_1.R
\name{my_fun_1}
\alias{my_fun_1}
\title{Esta función coge un df y calcula la media de cada columna}
\usage{
my_fun_1(df, print = FALSE)
}
\arguments{
\item{df}{Un data.frame}
}
\value{
Un nuevo df con la media de cada columna del df original \code{df}
}
\description{
Esta función coge un df y calcula la media de cada columna
}
\examples{
my_fun_1(mtcars)
}

