% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fun_2.R
\name{my_fun_2}
\alias{my_fun_2}
\title{Esta función suma 2 números}
\usage{
my_fun_2(x, y)
}
\arguments{
\item{x}{Un número}

\item{y}{Un número}
}
\value{
La suma de \code{x} e \code{y}
}
\description{
Esta función suma 2 números
}
\examples{
my_fun_2(1, 1)
my_fun_2(10, 1)
}

